requirements.txt
fastapi==0.110.2
paho_mqtt==1.6.1
python-socketio==5.11.2
uvicorn==0.29.0
